Moved 'Start Tour' button to the header and applied I18n translation
